Extract recipe run artifact finalizers
